requirements.txt
pyyaml
matplotlib
kubernetes
